Allow multiple cycles of execution per frame

+++ src/chip8.h
@@ -14,7 +14,7 @@ typedef struct Chip8
     uint8_t sp;
     uint8_t delay_timer;
     uint8_t sound_timer;
-    uint8_t speed;
+    uint32_t speed;
     bool paused;
 } Chip8;
